requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
decorator==5.1.1
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
executing==0.8.3
importlib-metadata==4.11.4
iniconfig==1.1.1
ipython==8.4.0
jedi==0.18.1
Markdown==3.3.7
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.2.2.post1
wcwidth==0.2.5
zipp==3.8.0
